CalDepGUIFuentes/requirements.txt
tk
Pillow
minizinc